Revert BigQueryPartition file relocation
