+++ flink-connector-bigquery/src/main/java/com/google/cloud/flink/bigquery/services/TablePartitionInfo.java
@@ -21,7 +21,7 @@ import org.apache.flink.annotation.Internal;
 import org.apache.flink.shaded.guava30.com.google.common.collect.Lists;
 
 import com.google.cloud.bigquery.StandardSQLTypeName;
-import com.google.cloud.flink.bigquery.common.utils.BigQueryPartition.PartitionType;
+import com.google.cloud.flink.bigquery.table.restrictions.BigQueryPartition.PartitionType;
 
 import java.time.Instant;
 import java.util.List;
