+++ flink-connector-bigquery/src/test/java/com/google/cloud/flink/bigquery/table/restrictions/BigQueryPartitionTest.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.flink.bigquery.common.utils;
+package com.google.cloud.flink.bigquery.table.restrictions;
 
 import org.apache.flink.shaded.guava30.com.google.common.collect.Lists;
